Enum vna::CliSubcommand

enum CliSubcommand {
    DataSync {
        scrape_interval: Option<u32>,
        max_news: u64,
        ingest_batch: NonZeroU32,
        n_shards: NonZeroU32,
        n_replicas: u32,
        leave_old_index: bool,
        data_source: DataSourceArgs,
    },
    Stats(Stats),
    Search {
        query: NonHollowString,
        field_name: Option<String>,
    },
    Snapshots(Snapshots),
}

Variants

DataSync

Run data synchronization job that will use the external data source

Fields of DataSync

scrape_interval: Option<u32>

Number of milliseconds to wait between successive scrapes of the external APIs. If none is specified the job is run only one time and shutdown right away, otherwise it will sleep for the specified amount of time and do the scraping inifinitely.

max_news: u64

Maximum amount of news to retain in Elasticsearch database

ingest_batch: NonZeroU32

Maximum size of the batch to ingest data with into Elasticsearch

n_shards: NonZeroU32

Number of shards to use for the Elasticsearch indices (min: 1)

n_replicas: u32

Number of relicas to create for the Elasticsearch indices

leave_old_index: bool

Whether to delete old index once the new one is created,

data_source: DataSourceArgs

Stats(Stats)

View varios statistics about the news via SVG charts

Search

Issue a fulltext search thru all the news

Fields of Search

query: NonHollowString

String of text to search for in Elasticsearch

field_name: Option<String>

Particular name of the field to search by in elasticsearch. If none is specified (which is the default) searches by all text fields

Snapshots(Snapshots)

Elasticsearch snapshots management commands